feat: permission checks for stage instance delete

+++ classes/MusicHandler.js
@@ -1,4 +1,4 @@
-const { MessageEmbed } = require('discord.js');
+const { MessageEmbed, Permissions } = require('discord.js');
 const { guildData, logger } = require('../shared.js');
 const { getLocale } = require('../functions.js');
 const { defaultLocale, defaultColor } = require('../settings.json');
@@ -17,6 +17,15 @@ module.exports = class MusicHandler {
 		this.player.disconnect();
 		bot.music.destroyPlayer(this.player.guildId);
 		if (voiceChannel?.type === 'GUILD_STAGE_VOICE') {
+			const voicePerms = bot.guilds.cache.get(this.player.guildId).channels.cache.get(this.player.channelId).permissionsFor(bot.user.id);
+			if (!voicePerms.has(['VIEW_CHANNEL', 'CONNECT', 'SPEAK'])) {
+				await this.player.musicHandler.locale('DISCORD_BOT_MISSING_PERMISSIONS_BASIC');
+				return;
+			}
+			if (!voicePerms.has(Permissions.STAGE_MODERATOR)) {
+				await this.player.musicHandler.locale('DISCORD_BOT_MISSING_PERMISSIONS_STAGE');
+				return;
+			}
 			if (voiceChannel.stageInstance?.topic) {
 				try {
 					await voiceChannel.stageInstance.delete();
@@ -28,6 +37,15 @@ module.exports = class MusicHandler {
 			return;
 		}
 		if (oldVoiceChannel.type === 'GUILD_STAGE_VOICE') {
+			const oldVoicePerms = bot.guilds.cache.get(this.player.guildId).channels.cache.get(oldVoiceChannel.id).permissionsFor(bot.user.id);
+			if (!oldVoicePerms.has(['VIEW_CHANNEL', 'CONNECT', 'SPEAK'])) {
+				await this.player.musicHandler.locale('DISCORD_BOT_MISSING_PERMISSIONS_BASIC');
+				return;
+			}
+			if (!oldVoicePerms.has(Permissions.STAGE_MODERATOR)) {
+				await this.player.musicHandler.locale('DISCORD_BOT_MISSING_PERMISSIONS_STAGE');
+				return;
+			}
 			if (oldVoiceChannel.stageInstance?.topic) {
 				try {
 					await oldVoiceChannel.stageInstance.delete();
